requirements.txt
google
pathlib
openai
google.genai
google-generativeai>=0.8.0
python-dotenv>=1.0.0
httpx>=0.27.0
flask>=3.0.0
gunicorn>=21.2.0
flask-sqlalchemy>=3.1.0
psycopg2-binary>=2.9.0